+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file. See [Conven
 
 ### Bug Fixes
 
-* **docs:** correct timeframe count and fix documentation issues ([15b2886](https://github.com/terrylica/gapless-crypto-clickhouse/commit/15b28860f716b122d9722363d230a770fb728172))
+- **docs:** correct timeframe count and fix documentation issues ([15b2886](https://github.com/terrylica/gapless-crypto-clickhouse/commit/15b28860f716b122d9722363d230a770fb728172))
 
 ## [8.0.0](https://github.com/terrylica/gapless-crypto-clickhouse/compare/v7.1.0...v8.0.0) (2025-11-22)